# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "77.240.39.57",
    "basePath": "/ai",
    "paths": {
...
        "/quiz/submit/{quizId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "quizId",
                        "name": "quizId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Submission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.Properties"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ErrorDetail"
                    }
                }
            }
        },
...
        "models.Properties": {
            "type": "object",
            "properties": {
                "additionalProp1": {
                    "type": "string"
                },
                "additionalProp2": {
                    "type": "string"
                },
                "additionalProp3": {
                    "type": "string"
                }
            }
        },
...
        "models.Submission": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/models.Properties"
                },
                "result": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}